PullRequest.comment_count lies
The PullRequest.comment_count property does not match the total comments
recorded for several pull requests. For instance, [1] reports 0 comments
but [2] lists 10 comments.

This behaviour does not seem tied to the state of the PR, where or by
whom the comment was made, or whether a comment is deleted. Indeed, it
seems entirely arbitrary, and although most counts are accurate there is
a non-trivial amount of disparities of non-trivial amounts.

Manually count not-deleted comments as a work-around.

[1] https://api.bitbucket.org/2.0/repositories/chromiumembedded/cef/pullrequests/17
[2] https://api.bitbucket.org/2.0/repositories/chromiumembedded/cef/pullrequests/17/comments