Ignore new API resource properties
Judging by commit 02c186a ("Fix API integration", 2019-08-11), Bitbucket
does not consider adding a new property to an existing resource a
backwards compatibility-breaking change (fairly). Reflect that in the
API integration. To prove it works, remove the new property
pullrequests.links.diffstat, which we didn't care about to begin with,
that broke the integration.