Report Time To First Approval, Feedback
The TTFC report is a decent indication of how long it takes for somebody
to react to a PR, as long as their reaction is to comment. If the
reviewer merely approves without commenting, the feedback period can be
much shorter than TTFC indicates. With approval activity we can also
report Time To First Approval, serving the same role for approvals
instead of comments, and Time To First Feedback, as the lower (better)
of the two.