Exclude self-feedback
It is quite common for a PR author to leave comments on their own PR. If
they happen to be the first person to do so, this will count as
"feedback", but it is not at least in any useful sense: although an
author may "think aloud" in a comment and produce refinements based on
that, it is not until somebody else has responded that the PR can be
said to have been reviewed and received feedback. The same argument
applies to approvals. Consequently, ignore all PR "feedback" produced by
the PR author.

This makes the long-tail of TTxy reports shorter and thicker; that is,
the proportional distribution shifts towards slower turn-around,
suggesting that quite a lot of "feedback" came from the author .