Excluded deleted comments from reports
As with the unresolved review interface, deleted comments probably
shouldn't count. Only the comment author can delete a comment, and
nobody can react to a deleted comment.

Don't apply the same filter to approval versus unapproval events.
Firstly, we don't have unapprovals so we can't. Secondly, approvals
afford a particular action (merging) that a comment does not in general.
This means that the only unapproval that invalidates a former approval
is one that follows immediately after an accidental approval -- any
other unapproval the PR author may have had time to react to.