Report duration density estimates
The "/duration/density" report is really a single UI entrypoint to
multiple existing reports. It draws a kernel density estimate of the
reported durations for the selected report, optionally with a comparison
time period. More than anything this was an exercise in playing around
with d3.js in a way that lets me compare different types of durations.

This is a massive commit because breaking it down at this point becomes
artificial and I can't be bothered.

A lot of effort went into getting a front-end that

- works with the back-end;
- works with obtuse JS module loading; and
- can be built from TypeScript.

I didn't want to build a separate application for the front-end (FE).
That would make it easier to build the FE in isolation but I suspect it
would make the whole application harder to build and distribute. That
said, Spring lets you load static assets from the file system, so it
seems like it could be done.

I built the first satisfactorily working version entirely in JS. This
convinced me that rewriting it in TS would probably be worthwhile.
Although getting to the point where everything fit together likely
shortened by life expectancy by several months, TS did turn out to speed
up later iterations. I was able to get SystemJS and core modules working
but in the end I went with a solution where the Main module can be
loaded by any page and the individual pages are responsible for
selecting the desired runtime. This causes some bloat from the
perspective of individual pages but simplifies maintenance.

The FE has to be built with "yarn bundle" after every change and the
compiled assets committed. This is potentially error prone. It uses a
lightweight bundler that does OK for this very specific purpose.

Another downside of the bundling setup is that "noEmit" must be false
(that is, "do emit"). The bundler produces no output without this, but
VS Code also produces output with it and it really shouldn't. They
cannot be disassociated.

I don't include the sourcemap files, not really for any reason other
than them just being noise.