Include approval count in /unresolved
I originally anticipated that an approved PR would transition to
resolved with relative ease and speed, and so including approval status
would be meaningless; and that comment and task activity -- or lack
thereof -- would indicate a promising venture for an enterprising
reviewer. Just like with commit's 26ccc83 ("Report Time To First
Approval, Feedback", 2020-02-01) TTFA report, that information is not
enough: in the instance where reviewers leave no other feedback than
approvals the report does not indicate that the author may be able to
resume work. Further, not every approval is born equal; some simply are
not representative, and with a count it's easier to detect activity.

This makes it more problematic that unapprovals are ignored, however.