Upgrade to Spring Boot 2.2.5
This is primarily motivated by an upcoming dependency on
HeaderResultMatchers::dateValue, which has a hard dependency on JUnit 4,
fixed in v5.1.8 and v5.2.4 [1]. It was either reintroduce JUnit 4 until
this inevitable upgrade, or upgrade now. We'll be in a better position
for Spring Boot 2.3, at least.

Other exciting changes include changing Freemarker view extensions from
.ftl to .ftlh [2], trivial API churn, new Flyway, and new PostgreSQL
JDBC driver.

[1] spring-projects/spring-framework@233b225b4fcb628e0a15c18f94941e2bdd5c6af2
[2] spring-projects/spring-boot#15131