Factor out TestIngester
A TestIngester is an Ingester that ingests the specified resource in a
blocking manner, for use in tests. An actual "blocking ingester" or
"direct ingester" passed to BackgroundIngester is a cleaner solution
that doesn't require a dedicated test helper class but then the retry
mechanism has to be rewritten as well. This is easier.